+++ Request/Payload.php
@@ -18,14 +18,20 @@ class Payload
 {
     use StdLibTrait;
 
-    private $_payloadBag;
+    private $payloadBag;
 
     /**
      * Constructor.
      */
     public function __construct()
     {
-        $this->_payloadBag = $this->arr(json_decode(file_get_contents("php://input"), true));
+        $data = json_decode(file_get_contents("php://input"), true);
+        $phpInput = file_get_contents('php://input');
+        if(is_null($data) && $phpInput != ''){
+            parse_str($phpInput, $data);
+        }
+
+        $this->payloadBag = $this->arr($data);
     }
 
     /**
@@ -38,7 +44,7 @@ class Payload
      */
     public function get($key, $value = null)
     {
-        return $this->_payloadBag->key($key, $value, true);
+        return $this->payloadBag->key($key, $value, true);
     }
 
     /**
@@ -48,6 +54,6 @@ class Payload
      */
     public function getAll()
     {
-        return $this->_payloadBag->val();
+        return $this->payloadBag->val();
     }
 }