+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 CHANGELOG
 =========
 
+1.4
+---------
+* Option to mock a Request method
+* Option to mock current url
+
+
+1.3
+---------
+* Removed the default Cookie prefix
+* Removed the default Session prefix
+
+
+1.2
+---------
+
+
 1.0.0-RC1
 ---------
 * Initial version